Almost everything according to norm
